fixing request a quote issues on fb

+++ image_collections/images.php
@@ -139,7 +139,7 @@ if(empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUE
 		|--------------------------------------------------------------------------
 		*/
 	    public function _save($raw,$mode = 'w+') {
-	    	if(file_exists($this->local)){
+	    	if(file_exists($this->local) && file_exists($this->data->save_as)){
 		        unlink($this->data->save_as);
 		    }
 		    file_put_contents($this->local, $raw);
